export class object builder
